[tests] closes #610890 - add a ground export test (known to FAIL).
This test triggers an assertion failure : <<Trying to build a plane
from NaN values !>>. The test is fine and the problem actually lie in
the changeset 673b01beebb3 <<do not throw exception for invalid
altitudes>>.

Indeed this changeset was too permissive, allowing to much invalid
data to propagate into the application. Next commit will fix that.