[python, tests, cmake] configure path for python test
Depending of the build configuration tested (Debug or Release) the
sys.path to be used must either include ``cython_d`` or ``cython``
respectively.

This require, first, for CMake to ask CTest to pass the build
configuration name to the test command and, second, for the python
test script to adjust their sys.path based on this argument.