[TYCalcul, TYSolverInterface] Limit dependency to solver data models
Use a PImpl-like idiom so that forward declaration for
AcousticProblemModel and AcousticResultModel are enough in the
headers and that their #inclusion is only required in the .cpp
files.

Please refer to the following resources for more background:

* http://mbevin.wordpress.com/2012/11/18/smart-pointers/
* http://stackoverflow.com/questions/6012157/is-stdunique-ptrt-required-to-know-the-full-definition-of-t

Please also note how Cython wrappings are impacted.
There is no specific support for unique_ptr so we break the use
of smart pointer by passing a real pointer T* obtained from
std::unique_ptr<T>::get().