[doc] Remove useless directory developer
related to #1649921

Remove the extra directory level `developer` which was a left-over
of having both the users and the developers documentation in the
repository. The users was finally kept as a word processing document
but the layout of the repository was not simplified accordingly.

Also rename instalation.rst to installing.rst to keep coherence
with the tasks oreinted help topics.