[tests] Add tests for matrix by point and matrix by vector multiplication
Currently there is no "operator *" defined to multiply a OMatrix by a
OPoint3D. "matrix * point" actually triggers an implicit conversion of the point
to a vector object and thus the operation "matrix * vector".

Yet, the "operator *(OVector3D)" defined in OMatrix class is not correct for a
vector, but is correct for a point (no translation on a vector)

So:
- test matrix_point_multiplication passes but it shouldn't since it actually
  triggers the execution of the multiplication matrix*vector.

- test matrix_vector_multiplication fails because the multiplication
  matrix*vector is written as a multiplication matrix*point.

Will be corrected in the next cset.


related to #1748363