[altimetry] Clean the handling of information attached to geometry.
Separated between information attached to input constraints and
points from other information wich may be added but are out of
scope of the responsibility of this class.

Simply exposes a clear_caches method in order to reset cached
information when the geometry wiil be modified : typically
information can not be realiabily attached to faces while all
inputs have not been done, and it must be reset when anyu input is
done.

related to #1961496