[altimetry] Add support for elevation meshes
An elevation mesh is a 2D mesh in whith each vertex has an
associated altitude. This requires to mark intersection of
constraints with a special value for unspecified altitudes.

A reference elevation mesh force all of its constraint to have an
altitude specfied and enable to compute the altitude for
intersection of constraints my merging this information.

A reference elevation mesh can be copied into an elevation mesh
and then be used to compute altitude for other types of
constraints inserted into the simple elevation mesh.

related to #1961496