[altimetry, cmake] Clean and fix python modules handling
The relative path for python components is now in the variable
TYMPAN_PythonModules set in the top level CMakeLists and used both
by the python sub-tree and the Tympan sub-tree of the sources. The
later uses it to configure the path to the scripts relative to the
installation dir used by the C++ code to invoke scripts.

The python sub-tree no longer is a CMake sub-project, thus
avoiding lots of ill-documented subtleties between root and
project BINARY/SOURCE_DIR.

The cython/python tree is now built in two (Release and Debug),
duplicating the pure python scripts, in the root build directory's
cython and cython_d directories.

altimetry modules and tests are fixed to search into the tympan
package.