[altimetry] don't keep UNSPECIFIED_ALTITUDES when updating mesh vertices altitude
When points are outside the convex hull of the triangulation, they are kept
but there altitude is set to UNSPECIFIED_ALTITUDE (see ReferenceElevationMesh.point_altitude).
This leads to an exception raised in ElevationMesh.point3d_for_vertex).
(see for example with project
tests/data/projects-panel/TEST_DIFFRACTION_BUTTE_CARTO_NO_RESU.xml).

For now, just set the altitude of these points to 0.

Related to #2741286