[GUI] ask for an altimetry update on the TYSiteFrame instead of on its TYSiteNode
The need for an altimetry update is mainly triggered when an element is added,
deleted or modified in the GUI.
The button allowing to trigger the altimetry recomputation must be enabled or
disabled depending on the altimetry status. It is consequently easier to put
this data on the GUI side.

Related to #2741286