[altimetry] Correct level curves in some XML projects used in the tests
Level curves must be stricly included in a landtake.
Remove the ones that were defined on a landtake and that intersect it at some
points. Instead, activate the option allowing to use the landtake as a level
curve (with an altitude equal to the one of the late level curve),
which is equivalent.


Related to #2741286