[TYPoint] remove call to operator = in TYPoint constructor
As TYPoint has no proper data members, the only effect of this affectation is
to reset the object UUID.

It is not the desired effect since an UUID has already been affected to the
object during call to TYElement constructor and there is no reason why
modifiying it. The position is set to the TYPoint during call to OPoint3D
constructor.

Related to #2810138