Fix solver parameters under windows
As the solvers are compiled into static libraries under Windows, even if the
solver parameters are set from the cython code, a call to SolverConfiguration::get() in the C++ code
always leads to the creation of a second SolverConfiguration instance, with default
values.
The best solution would be to compile the solvers as dynamic libraries under
Windows (as it is done under linux).

But for now a quick fix is to give a pointer to these parameters through the solve()
method (as it is already done for the solver problem and result objects).

closes #2921897