[altimetry] Replace the material face flooder by a simpler geometric algorithm
This algorithm just walks through all faces of the mesh, computes a point
inside and then finds the geometric feature (within material areas) that point
belongs to.
This simple algorithm appears to solve the issue of association of materials
with mesh faces described in #2875340. So the failing functional test now
passes.

Remove the MaterialFaceFlooder class and respective tests.

Closes #3176147.