[cmake] move resources dir to the root directory instead of the tests directory
Because its content is not related to the tests, and because it should always located in the same place (in the code dir, in the build dir and in the install dir).

This will allow to locate solver param json file both from the tests and from the application (see next cset).