Join all polygons of the faces on ground into an union in TYSiteNode::groundBasedFaces
For infrastructure volumes (such as TYEcran) having more than one face on
ground, groundBasedFaces only returned the contour of the last visited face,
which resulted in respective volumes to be partially accounted for in the
altimetry processing.

Now accumulate the polygons of all faces on ground of a volume and compute the
union using `join_polygons` function introduced in cgal_bridge module.

(Bitbucket fix #3)