[cython] correct assert syntax, and fix the bug this wrong syntax was hiding
The OTriangle constructor that takes 3 indices sets the triangle points coordinates to NaN.
The assert checks that the points of "pts" corresponding to the indices stored in the newly created triangles are the same as the points _A, _B and _C of the triangles (which were set to NaN).

Here we set them to there right values immediatly after building the triangles.

It isn't a very intuitive way of building a triangle, and OTriangle design should be improved: see
issue #29 in bitbucket.