Add road constraint feature
We insert roads (polyedges) into altimetry for the mesh calculation.
We treat roads as level curves except for altitudes merging.
We must use the type of vertex (Road or otherwise) to allow only roads
to have points of different altitudes during the merge.
Moreover, unlike a classic polyline, a polyedge can't be closed and
its points are always connected.