Take the inclination angles of the road in account
The angles of the road makes it possible to determine the altitudes of
each track relative to the main line of the road.
We need to add two more lines in the definition of the Road object
that are very close and parallel to the main line of the road but at
different altitudes.